requirements.txt
google-cloud-bigquery>=3.12,<4
pandas>=2.2,<3
snowflake-connector-python[pandas]>=3.7,<4
dbt-snowflake>=1.8,<1.9
python-dotenv>=1.0,<2